refactor tacotron2 loss tests
